Start of major restructuring.
