+++ main.h
@@ -1,0 +1,38 @@
+#ifndef _MAIN_H_
+#define _MAIN_H_
+
+#define hidecursor gotoxy(1, 1)
+
+/**
+ * Displays confirmation message on bottom line of screen and waits for user to
+ * type 'y' or 'n'. The line is erased once either choice has been made. Return
+ * values are non-zero for 'yes' and zero for 'no'.
+ */
+int confirm_yn(char *message);
+
+/**
+ * Move the cursor to the nth freecell.
+ */
+void goto_freecell(int index);
+
+void gotocc(char cascadei, char cardi);
+
+void pretty_borders(char x, char y);
+
+void refresh(void);
+
+void refresh_freecells(void);
+
+void refresh_homecells(void);
+
+/**
+ * Initializes static variables and sets up display. Only call once!
+ */
+void init(void);
+
+/**
+ * Program point of entry.
+ */
+int main(void);
+
+#endif
